+++ md/state.c
@@ -1,5 +1,5 @@
 #include "state.h"
 
-volatile int64_t counter = 0;
+volatile int32_t counter = 0;
 volatile int8_t power = 0;
 volatile float pid_target = 0;
