+++ md/state.h
@@ -4,7 +4,7 @@
 #include "Arduino.h"
 
 // encoder value
-extern volatile int64_t counter;
+extern volatile int32_t counter;
 // -100 <= power <= 100
 extern volatile int8_t power;
 extern volatile float pid_target;
